+++ CHANGELOG.md
@@ -3,7 +3,12 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
-[//]: # (## [Unreleased])
+## [Unreleased]
+
+### Fixed
+
+- [#55](https://github.com/oleasteo/koreader-screenlockpin/issues/55)
+  Updater compatibility with koreader 2026.07.
 
 ## [2026.09] - 2026-09-12